pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibemouse"
version = "0.1.0"
description = "Mouse side-button voice dictation with SenseVoice"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
]
authors = [{ name = "VibeMouse" }]
dependencies = [
  "numpy>=1.24,<2",
  "sounddevice>=0.4.6",
  "pynput>=1.7.7",
  "evdev>=1.7.1; sys_platform == 'linux'",
  "PyGObject>=3.46.0; sys_platform == 'linux'",
  "pystray>=0.19.5; sys_platform == 'win32'",
  "Pillow>=10.0; sys_platform == 'win32'",
  "sherpa-onnx>=1.10.0",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0"]

[project.scripts]
vibemouse = "vibemouse.main:main"

[project.gui-scripts]
vibemouse-gui = "vibemouse.main:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["vibemouse*"]
